Fix install scripts
