+++ torchbenchmark/models/lit_llama_lora/install.py
@@ -1,4 +1,4 @@
-from ..lit_llama.install import install_lit_llama
+from torchbenchmark.util.framework.lit_llama import install_lit_llama
 
 if __name__ == '__main__':
     install_lit_llama()
